Analyses

Documentation

type DifTrans m = (Map Int (m -> m), Map Int (m -> m -> m))

Transfer functions per label, grouped by unary and binary

insertL :: Int -> (m -> m) -> DifTrans m -> DifTrans m

Insert a unary transfer function at the label

insertR :: Int -> (m -> m -> m) -> DifTrans m -> DifTrans m

Insert a binary transfer function at the label

singleL :: Int -> (m -> m) -> DifTrans m

Create a singleton Map with this unary transfer function at the label

singleR :: Int -> (m -> m -> m) -> DifTrans m

Create a singleton Map with this binary transfer function at the label

lookupL :: Int -> DifTrans p -> p -> p

Lookup a unary transfer function

lookupR :: Int -> DifTrans p -> p -> p -> p

Lookup a binary transfer function

data Proc''

Internal representation of "Proc'" as a record, for convenience

Constructors

Proc''

Fields

procEntry :: Int

procExit :: Int

procName :: String

procInp :: [String]

procOut :: String

type DStar = [(String, Proc'')]

Type of procedure lookup tables

type Edge = (Int, Int)

data Inter

Constructors

Inter Int Int Int Int

findProc :: String -> DStar -> Proc''

Unsafely lookup a procedure by its name

groupBy :: Ord k => (a -> k) -> [a] -> Map k [a]

Group by a key function

groupBy' :: Ord k => (a -> (k, b)) -> [a] -> Map k [b]

Group by the keys of a key/value function

concatSnd :: [(k, [(m, v)])] -> [(k, (m, v))]

flipMap :: (Ord k, Ord m) => Map k (Map m v) -> Map m (Map k v)

Flip the inner and outer keys of a nested map