AnalysesConversion

Documentation

data Dir

Constructors

Forward

Backward

data Flow

Constructors

Flow

Fields

initial :: Int

finals :: Set Int

edges :: Set Edge

interflow :: Set Inter

data Analysis p

Constructors

Analysis

Fields

direction :: Dir

difTrans :: DifTrans p

extremal :: p

swap :: (a, b) -> (b, a)

unpackFlow :: Bool -> Flow -> (Set Int, Set Edge, Set Inter)

Unpack a flow into a tuple of extremals, edges, and interflow edges. Reverses the necessary components for backwards flows.

group :: (Ord a, Ord b) => [(a, b)] -> Map a (Set b)

Equivalent of python's groupby: for each fst element occuring in the list, gather all snd's belonging to tuples with matching fst into a list

analysisToFramework :: Analysis p -> Flow -> MonotoneFramework p

Combine an analysis and a flow into a monotone framework