AttributeGrammar

Documentation

indent :: [String] -> [String]

showLabel :: Int -> String

addSemicolon :: [String] -> [String]

parensIf :: Bool -> String -> String

data BExpr

Constructors

BConst Bool

BVar String

LessThan IExpr IExpr

GreaterThan IExpr IExpr

LessEqual IExpr IExpr

GreaterEqual IExpr IExpr

IEqual IExpr IExpr

BEqual BExpr BExpr

And BExpr BExpr

Or BExpr BExpr

Not BExpr

sem_BExpr :: BExpr -> T_BExpr

type T_BExpr = (ConstEnv -> Maybe ConstLat, Set String, Int, String, BExpr)

data Inh_BExpr

Constructors

Inh_BExpr

data Syn_BExpr

Constructors

Syn_BExpr

Fields

constantPropagation_Syn_BExpr :: ConstEnv -> Maybe ConstLat

freeVars_Syn_BExpr :: Set String

precedence_Syn_BExpr :: Int

pretty_Syn_BExpr :: String

self_Syn_BExpr :: BExpr

wrap_BExpr :: T_BExpr -> Inh_BExpr -> Syn_BExpr

sem_BExpr_BConst :: Bool -> T_BExpr

sem_BExpr_BVar :: String -> T_BExpr

sem_BExpr_LessThan :: T_IExpr -> T_IExpr -> T_BExpr

sem_BExpr_GreaterThan :: T_IExpr -> T_IExpr -> T_BExpr

sem_BExpr_LessEqual :: T_IExpr -> T_IExpr -> T_BExpr

sem_BExpr_GreaterEqual :: T_IExpr -> T_IExpr -> T_BExpr

sem_BExpr_IEqual :: T_IExpr -> T_IExpr -> T_BExpr

sem_BExpr_BEqual :: T_BExpr -> T_BExpr -> T_BExpr

sem_BExpr_And :: T_BExpr -> T_BExpr -> T_BExpr

sem_BExpr_Or :: T_BExpr -> T_BExpr -> T_BExpr

sem_BExpr_Not :: T_BExpr -> T_BExpr

data Expr

Constructors

B BExpr

I IExpr

sem_Expr :: Expr -> T_Expr

type T_Expr = (ConstEnv -> Maybe ConstLat, Set String, String, Expr)

data Inh_Expr

Constructors

Inh_Expr

data Syn_Expr

Constructors

Syn_Expr

Fields

constantPropagation_Syn_Expr :: ConstEnv -> Maybe ConstLat

freeVars_Syn_Expr :: Set String

pretty_Syn_Expr :: String

self_Syn_Expr :: Expr

wrap_Expr :: T_Expr -> Inh_Expr -> Syn_Expr

sem_Expr_B :: T_BExpr -> T_Expr

sem_Expr_I :: T_IExpr -> T_Expr

type Exprs = [Expr]

sem_Exprs :: Exprs -> T_Exprs

type T_Exprs = ([ConstEnv -> Maybe ConstLat], [Set String], String, Exprs)

data Inh_Exprs

Constructors

Inh_Exprs

data Syn_Exprs

Constructors

Syn_Exprs

Fields

constantPropagation_Syn_Exprs :: [ConstEnv -> Maybe ConstLat]

freeVars_Syn_Exprs :: [Set String]

pretty_Syn_Exprs :: String

self_Syn_Exprs :: Exprs

wrap_Exprs :: T_Exprs -> Inh_Exprs -> Syn_Exprs

sem_Exprs_Cons :: T_Expr -> T_Exprs -> T_Exprs

sem_Exprs_Nil :: T_Exprs

data IExpr

Constructors

IConst Int

Var String

Plus IExpr IExpr

Minus IExpr IExpr

Times IExpr IExpr

Divide IExpr IExpr

Deref IExpr

sem_IExpr :: IExpr -> T_IExpr

type T_IExpr = (ConstEnv -> Maybe ConstLat, Set String, Int, String, IExpr)

data Inh_IExpr

Constructors

Inh_IExpr

data Syn_IExpr

Constructors

Syn_IExpr

Fields

constantPropagation_Syn_IExpr :: ConstEnv -> Maybe ConstLat

freeVars_Syn_IExpr :: Set String

precedence_Syn_IExpr :: Int

pretty_Syn_IExpr :: String

self_Syn_IExpr :: IExpr

wrap_IExpr :: T_IExpr -> Inh_IExpr -> Syn_IExpr

sem_IExpr_IConst :: Int -> T_IExpr

sem_IExpr_Var :: String -> T_IExpr

sem_IExpr_Plus :: T_IExpr -> T_IExpr -> T_IExpr

sem_IExpr_Minus :: T_IExpr -> T_IExpr -> T_IExpr

sem_IExpr_Times :: T_IExpr -> T_IExpr -> T_IExpr

sem_IExpr_Divide :: T_IExpr -> T_IExpr -> T_IExpr

sem_IExpr_Deref :: T_IExpr -> T_IExpr

data Proc

Constructors

Proc String [String] String Stat

sem_Proc :: Proc -> T_Proc

type T_Proc = Int -> ((String, Proc''), Int, Proc', Proc)

data Inh_Proc

Constructors

Inh_Proc

Fields

label_Inh_Proc :: Int

data Syn_Proc

Constructors

Syn_Proc

Fields

dStar_Syn_Proc :: (String, Proc'')

label_Syn_Proc :: Int

labelled_Syn_Proc :: Proc'

self_Syn_Proc :: Proc

wrap_Proc :: T_Proc -> Inh_Proc -> Syn_Proc

sem_Proc_Proc :: String -> [String] -> String -> T_Stat -> T_Proc

data Proc'

Constructors

Proc' Int Int String [String] String Stat'

sem_Proc' :: Proc' -> T_Proc'

type T_Proc' = DStar -> (DifTrans ConstBranchLat, DifTrans PtConstLat, Set Int, Set Edge, Int, Set Inter, [String], Proc', DifTrans (Set String))

data Inh_Proc'

Constructors

Inh_Proc'

Fields

dStar_Inh_Proc' :: DStar

data Syn_Proc'

Constructors

Syn_Proc'

Fields

constBranchT_Syn_Proc' :: DifTrans ConstBranchLat

constantPropagation_Syn_Proc' :: DifTrans PtConstLat

final_Syn_Proc' :: Set Int

flow_Syn_Proc' :: Set Edge

init_Syn_Proc' :: Int

interflow_Syn_Proc' :: Set Inter

pretty_Syn_Proc' :: [String]

self_Syn_Proc' :: Proc'

strongLive_Syn_Proc' :: DifTrans (Set String)

wrap_Proc' :: T_Proc' -> Inh_Proc' -> Syn_Proc'

sem_Proc'_Proc' :: Int -> Int -> String -> [String] -> String -> T_Stat' -> T_Proc'

type Procs = [Proc]

sem_Procs :: Procs -> T_Procs

type T_Procs = Int -> (DStar, Int, Procs', Procs)

data Inh_Procs

Constructors

Inh_Procs

Fields

label_Inh_Procs :: Int

data Syn_Procs

Constructors

Syn_Procs

Fields

dStar_Syn_Procs :: DStar

label_Syn_Procs :: Int

labelled_Syn_Procs :: Procs'

self_Syn_Procs :: Procs

wrap_Procs :: T_Procs -> Inh_Procs -> Syn_Procs

sem_Procs_Cons :: T_Proc -> T_Procs -> T_Procs

sem_Procs_Nil :: T_Procs

type Procs' = [Proc']

sem_Procs' :: Procs' -> T_Procs'

type T_Procs' = DStar -> (DifTrans ConstBranchLat, DifTrans PtConstLat, Set Edge, Set Inter, [String], Procs', DifTrans (Set String))

data Inh_Procs'

Constructors

Inh_Procs'

Fields

dStar_Inh_Procs' :: DStar

data Syn_Procs'

Constructors

Syn_Procs'

Fields

constBranchT_Syn_Procs' :: DifTrans ConstBranchLat

constantPropagation_Syn_Procs' :: DifTrans PtConstLat

flow_Syn_Procs' :: Set Edge

interflow_Syn_Procs' :: Set Inter

pretty_Syn_Procs' :: [String]

self_Syn_Procs' :: Procs'

strongLive_Syn_Procs' :: DifTrans (Set String)

wrap_Procs' :: T_Procs' -> Inh_Procs' -> Syn_Procs'

sem_Procs'_Cons :: T_Proc' -> T_Procs' -> T_Procs'

sem_Procs'_Nil :: T_Procs'

data Program

Constructors

Program Procs Stat

sem_Program :: Program -> T_Program

type T_Program = (Program', Program)

data Inh_Program

Constructors

Inh_Program

data Syn_Program

Constructors

Syn_Program

Fields

labelled_Syn_Program :: Program'

self_Syn_Program :: Program

wrap_Program :: T_Program -> Inh_Program -> Syn_Program

sem_Program_Program :: T_Procs -> T_Stat -> T_Program

data Program'

Constructors

Program' Procs' Stat' DStar

sem_Program' :: Program' -> T_Program'

type T_Program' = (DifTrans ConstBranchLat, DifTrans PtConstLat, Set Int, Set Edge, Int, Set Inter, String, Program', DifTrans (Set String))

data Inh_Program'

Constructors

Inh_Program'

data Syn_Program'

Constructors

Syn_Program'

Fields

constBranchT_Syn_Program' :: DifTrans ConstBranchLat

constantPropagation_Syn_Program' :: DifTrans PtConstLat

final_Syn_Program' :: Set Int

flow_Syn_Program' :: Set Edge

init_Syn_Program' :: Int

interflow_Syn_Program' :: Set Inter

pretty_Syn_Program' :: String

self_Syn_Program' :: Program'

strongLive_Syn_Program' :: DifTrans (Set String)

wrap_Program' :: T_Program' -> Inh_Program' -> Syn_Program'

sem_Program'_Program' :: T_Procs' -> T_Stat' -> DStar -> T_Program'

data Stat

Constructors

Skip

IfThenElse BExpr Stat Stat

While BExpr Stat

Print Expr

Call String Exprs String

IAssign String IExpr

BAssign String BExpr

Seq Stat Stat

Malloc String IExpr

Free IExpr

RefAssign IExpr IExpr

Continue

Break

sem_Stat :: Stat -> T_Stat

type T_Stat = Int -> (Int, Stat', Stat)

data Inh_Stat

Constructors

Inh_Stat

Fields

label_Inh_Stat :: Int

data Syn_Stat

Constructors

Syn_Stat

Fields

label_Syn_Stat :: Int

labelled_Syn_Stat :: Stat'

self_Syn_Stat :: Stat

wrap_Stat :: T_Stat -> Inh_Stat -> Syn_Stat

sem_Stat_Skip :: T_Stat

sem_Stat_IfThenElse :: BExpr -> T_Stat -> T_Stat -> T_Stat

sem_Stat_While :: BExpr -> T_Stat -> T_Stat

sem_Stat_Print :: Expr -> T_Stat

sem_Stat_Call :: String -> Exprs -> String -> T_Stat

sem_Stat_IAssign :: String -> IExpr -> T_Stat

sem_Stat_BAssign :: String -> BExpr -> T_Stat

sem_Stat_Seq :: T_Stat -> T_Stat -> T_Stat

sem_Stat_Malloc :: String -> IExpr -> T_Stat

sem_Stat_Free :: IExpr -> T_Stat

sem_Stat_RefAssign :: IExpr -> IExpr -> T_Stat

sem_Stat_Continue :: T_Stat

sem_Stat_Break :: T_Stat

data Stat'

Constructors

Skip' Int

IfThenElse' Int BExpr Stat' Stat'

While' Int BExpr Stat'

Print' Int Expr

Call' Int Int String Exprs String

IAssign' Int String IExpr

BAssign' Int String BExpr

Seq' Stat' Stat'

Malloc' Int String IExpr

Free' Int IExpr

RefAssign' Int IExpr IExpr

Continue' Int

Break' Int

sem_Stat' :: Stat' -> T_Stat'

type T_Stat' = Maybe Int -> DStar -> Maybe Int -> (Set Int, DifTrans ConstBranchLat, DifTrans PtConstLat, Set Int, Set Edge, Int, Set Inter, Bool, Bool, [String], Stat', DifTrans (Set String))

data Inh_Stat'

Constructors

Inh_Stat'

Fields

continueLabel_Inh_Stat' :: Maybe Int

dStar_Inh_Stat' :: DStar

successor_Inh_Stat' :: Maybe Int

data Syn_Stat'

Constructors

Syn_Stat'

Fields

breakLabels_Syn_Stat' :: Set Int

constBranchT_Syn_Stat' :: DifTrans ConstBranchLat

constantPropagation_Syn_Stat' :: DifTrans PtConstLat

final_Syn_Stat' :: Set Int

flow_Syn_Stat' :: Set Edge

init_Syn_Stat' :: Int

interflow_Syn_Stat' :: Set Inter

isSingle_Syn_Stat' :: Bool

isSkip_Syn_Stat' :: Bool

pretty_Syn_Stat' :: [String]

self_Syn_Stat' :: Stat'

strongLive_Syn_Stat' :: DifTrans (Set String)

wrap_Stat' :: T_Stat' -> Inh_Stat' -> Syn_Stat'

sem_Stat'_Skip' :: Int -> T_Stat'

sem_Stat'_IfThenElse' :: Int -> T_BExpr -> T_Stat' -> T_Stat' -> T_Stat'

sem_Stat'_While' :: Int -> T_BExpr -> T_Stat' -> T_Stat'

sem_Stat'_Print' :: Int -> T_Expr -> T_Stat'

sem_Stat'_Call' :: Int -> Int -> String -> T_Exprs -> String -> T_Stat'

sem_Stat'_IAssign' :: Int -> String -> T_IExpr -> T_Stat'

sem_Stat'_BAssign' :: Int -> String -> T_BExpr -> T_Stat'

sem_Stat'_Seq' :: T_Stat' -> T_Stat' -> T_Stat'

sem_Stat'_Malloc' :: Int -> String -> T_IExpr -> T_Stat'

sem_Stat'_Free' :: Int -> T_IExpr -> T_Stat'

sem_Stat'_RefAssign' :: Int -> T_IExpr -> T_IExpr -> T_Stat'

sem_Stat'_Continue' :: Int -> T_Stat'

sem_Stat'_Break' :: Int -> T_Stat'