ConstantBranch

Documentation

newtype Intersect a

Newtype for sets with the intersection semigroup (as opposed to union).

Constructors

Intersect (Set a)

type ConstBranchLat = (PtConstLat, Maybe (Intersect Int))

Lattice for branch aware constant propagation

element :: Ord a => a -> Maybe (Intersect a) -> Bool

Element relation for sets with an universal set adjoined: everything is in the universe (Nothing)

ignoreDead :: Int -> (PtConstLat -> PtConstLat) -> ConstBranchLat -> ConstBranchLat

Ignore the unary transfer function at this label if it is dead

ignoreDead2 :: Int -> (PtConstLat -> PtConstLat -> PtConstLat) -> ConstBranchLat -> ConstBranchLat -> ConstBranchLat

Ignore the binary transfer function at this label if it is dead

constBranchId :: Int -> ConstBranchLat -> ConstBranchLat

Discard the analysis value at this label if it is dead

constBranchIf :: (ConstEnv -> Maybe ConstLat) -> Int -> Int -> ConstBranchLat -> ConstBranchLat

Transfer function for if: if the condition is always true, the else branch is unreachable (dead) and the then branch is made reachable, and vice versa when the condition is always false. Otherwise, both are reachable.

constBranchWhile :: Maybe Int -> (ConstEnv -> Maybe ConstLat) -> Int -> Int -> ConstBranchLat -> ConstBranchLat

Transfer function for while: if the condition is always false, the body is dead, and the label after the loop is made reachable, if it is always true, then the program loops, and the body is made reachable, and the label after is dead. Otherwise, both are reachable.

doIf :: (a -> b -> b) -> Maybe a -> b -> b

Apply a binary function with optional first argument, if Nothing, do nothing.