ConstantProp

Documentation

data ConstLat

Either Int Bool

Constructors

CI Int

CB Bool

newtype ConstEnv

Newtype for maps with the semigroup of intersection (as opposed to left-biased union)

Constructors

ConstEnv (Map String ConstLat)

type PtConstLat = Maybe ConstEnv

Lattice for the constant propagation analysis

constEmpty :: ConstEnv

setConst :: String -> Maybe ConstLat -> ConstEnv -> ConstEnv

Set this variable to a (non)-constant

setNonConst :: String -> ConstEnv -> ConstEnv

Mark this variable as non-constant

getConst :: String -> ConstEnv -> Maybe ConstLat

Lookup this variable in a ConstEnv

updateConst :: String -> (ConstEnv -> Maybe ConstLat) -> ConstEnv -> ConstEnv

Set this variable to the value of the given expression, as a function from environment to value, evaluated in the given context

callConst :: [String] -> String -> [ConstEnv -> Maybe ConstLat] -> ConstEnv -> ConstEnv

Transfer function for call: each parameter is initialized to the lattice value of its input expression

retConst :: String -> [String] -> String -> ConstEnv -> ConstEnv -> ConstEnv

Transfer function for return: restore all values of the parameters and the return parameter, and copy the value from the return parameter to the context before the call.

Functions lifting arithmetic/boolean operations to the lattice

cIII :: (Int -> Int -> Int) -> Maybe ConstLat -> Maybe ConstLat -> Maybe ConstLat

constMul :: Maybe ConstLat -> Maybe ConstLat -> Maybe ConstLat

cIIB :: (Int -> Int -> Bool) -> Maybe ConstLat -> Maybe ConstLat -> Maybe ConstLat

cBBB :: (Bool -> Bool -> Bool) -> Maybe ConstLat -> Maybe ConstLat -> Maybe ConstLat

constAnd :: Maybe ConstLat -> Maybe ConstLat -> Maybe ConstLat

constOr :: Maybe ConstLat -> Maybe ConstLat -> Maybe ConstLat

cBB :: (Bool -> Bool) -> Maybe ConstLat -> Maybe ConstLat