Latex

Documentation

latexPrint :: (Maybe propertySpace -> String) -> (Map Label (ContextSensitive propertySpace), Map Label (ContextSensitive propertySpace), [(Map Label (ContextSensitive propertySpace), [(Label, Label)])]) -> IO ()

Format and print an analysis result as a latex array

latexPrinter :: String -> (Maybe propertySpace -> String) -> Map Label (ContextSensitive propertySpace) -> String

Format a table of properties as a latex array, with the given name and property formatter

constantPropagationTex :: Maybe PtConstLat -> String

constEnvTex :: ConstEnv -> String

constantBranchTex :: Maybe ConstBranchLat -> String

setTex :: (a -> String) -> Set a -> String

strongLiveTex :: Maybe (Set String) -> String