MonotoneFrameworks

Documentation

class (Monoid m, Eq m) => BoundedSemiLattice m where

Minimal complete definition

Nothing

Methods

lessOrEquals :: m -> m -> Bool

bottom :: BoundedSemiLattice m => m

type Label = Int

data MonotoneFramework propertySpace

Constructors

MonotoneFramework

Fields

flow :: Map Label (Set Label)

map of labels to the set of destination labels of all outgoing flow from the first label

extremalLabels :: Set Label

extremalValue :: propertySpace

transferFunctions :: Map Label (propertySpace -> propertySpace)

interproceduralFragment :: InterproceduralFragment propertySpace

newtype InterproceduralFragment propertySpace

invariant. call label is functionally dependent on the return label

Constructors

InterproceduralFragment (Map Label (Label, propertySpace -> propertySpace -> propertySpace))

mfpSolution

Arguments

:: forall propertySpace. BoundedSemiLattice propertySpace

=> MonotoneFramework propertySpace
monotone framework

-> (Map Label propertySpace, Map Label propertySpace, [(Map Label propertySpace, [(Label, Label)])])
\(\mathrm{Analysis}_\circ\), \(\mathrm{Analysis}_\bullet\), and all intermediate steps

interprocedural MFP solution

warning. partial. transferFunctions must cover all labels except for call return labels.

warning. termination is only guaranteed if propertySpace satisfies the Ascending Chain Condition.

mapFunctionWithCallLabel :: forall p0 p1. (Label -> (p0 -> p0 -> p0) -> p1 -> p1 -> p1) -> InterproceduralFragment p0 -> InterproceduralFragment p1

lookupCall :: forall propertySpace. Label -> InterproceduralFragment propertySpace -> Maybe (Label, propertySpace -> propertySpace -> propertySpace)

lookupReturn :: Label -> InterproceduralFragment propertySpace -> Maybe (Label, propertySpace -> propertySpace -> propertySpace)