StrongLive

Documentation

moveAlive :: Set String -> String -> Set String -> Set String -> Set String

If name is live in src, then the set freeVars is live in dst

keepAlive :: String -> Set String -> Set String -> Set String

If name is live, then the set freeVars is live

aliveInCall :: [String] -> String -> [Set String] -> Set String -> Set String -> Set String

Transfer function for call: everything that lives in the procedure remains live, except for the parameters, everything that lives after the return remains live, except for the return parameter, and for each parameter that is live in the procedure, all the free variables of its input value are live.

aliveInExit :: [String] -> String -> String -> Set String -> Set String

Transfer function for return: everything that lives after the return remains live, except for the output variable and input parameters, if the output variable was live then the output parameter also becomes live

applyIf :: (a -> a) -> Bool -> a -> a

Orphan instances

BoundedSemiLattice (Set String)

Instance details

Methods

lessOrEquals :: Set String -> Set String -> Bool